More work on probability maps
